fix: review comments incorporation
