+++ tests/helpers/kube.go
@@ -2688,3 +2688,233 @@ func DeleteExistingNodeEvents(ctx context.Context, t *testing.T, c klient.Client
 
 	return nil
 }
+
+// ListDaemonSetPods returns all pods owned by the specified DaemonSet.
+func ListDaemonSetPods(ctx context.Context, client klient.Client, namespace, dsName string) ([]v1.Pod, error) {
+	var podList v1.PodList
+
+	err := client.Resources(namespace).List(ctx, &podList)
+	if err != nil {
+		return nil, err
+	}
+
+	var dsPods []v1.Pod
+
+	for _, pod := range podList.Items {
+		for _, ownerRef := range pod.OwnerReferences {
+			if ownerRef.Kind == "DaemonSet" && ownerRef.Name == dsName {
+				dsPods = append(dsPods, pod)
+
+				break
+			}
+		}
+	}
+
+	return dsPods, nil
+}
+
+// CleanupDaemonSet deletes a DaemonSet and waits for it and its pods to be fully deleted.
+func CleanupDaemonSet(ctx context.Context, t *testing.T, client klient.Client, namespace, name string) {
+	t.Helper()
+
+	ds := &appsv1.DaemonSet{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      name,
+			Namespace: namespace,
+		},
+	}
+
+	// Check if DaemonSet exists first - skip cleanup if not found
+	err := client.Resources(namespace).Get(ctx, name, namespace, ds)
+	if apierrors.IsNotFound(err) {
+		t.Logf("DaemonSet %s/%s does not exist, skipping cleanup", namespace, name)
+		return
+	}
+
+	if err != nil {
+		t.Logf("Warning: error checking DaemonSet existence: %v", err)
+	}
+
+	// Delete the DaemonSet
+	if err := client.Resources().Delete(ctx, ds); err != nil {
+		if apierrors.IsNotFound(err) {
+			t.Logf("DaemonSet %s/%s already deleted", namespace, name)
+			return
+		}
+
+		t.Logf("Note: DaemonSet deletion returned error: %v", err)
+	}
+
+	// Wait for pods to be deleted (short timeout since terminationGracePeriodSeconds=1)
+	require.Eventually(t, func() bool {
+		pods, err := ListDaemonSetPods(ctx, client, namespace, name)
+		return err == nil && len(pods) == 0
+	}, 30*time.Second, 1*time.Second, "timed out waiting for pods of DaemonSet %s/%s to be deleted", namespace, name)
+
+	// Wait for DaemonSet to be fully deleted
+	require.Eventually(t, func() bool {
+		err := client.Resources(namespace).Get(ctx, name, namespace, &appsv1.DaemonSet{})
+		if err == nil {
+			return false // Object still exists
+		}
+
+		if apierrors.IsNotFound(err) {
+			return true // Deleted successfully
+		}
+
+		// Transient error - log and retry
+		t.Logf("Note: transient error checking DaemonSet deletion: %v", err)
+
+		return false
+	}, 15*time.Second, 1*time.Second, "timed out waiting for DaemonSet %s/%s to be fully deleted", namespace, name)
+}
+
+// UpdateDaemonSet gets a DaemonSet, applies a modifier function, and updates it.
+func UpdateDaemonSet(
+	ctx context.Context, client klient.Client, namespace, name string, modifier func(*appsv1.DaemonSet),
+) error {
+	ds := &appsv1.DaemonSet{}
+
+	err := client.Resources(namespace).Get(ctx, name, namespace, ds)
+	if err != nil {
+		return err
+	}
+
+	modifier(ds)
+
+	return client.Resources().Update(ctx, ds)
+}
+
+// CheckPodCrashLoopBackOff checks if a pod has CrashLoopBackOff in its container statuses.
+// Returns true if CrashLoopBackOff is found.
+func CheckPodCrashLoopBackOff(statuses []v1.ContainerStatus) bool {
+	for _, cs := range statuses {
+		if cs.State.Waiting != nil && cs.State.Waiting.Reason == "CrashLoopBackOff" {
+			return true
+		}
+	}
+
+	return false
+}
+
+// LogContainerStatuses logs the current state of container statuses for debugging.
+func LogContainerStatuses(t *testing.T, statuses []v1.ContainerStatus, containerType string) {
+	t.Helper()
+
+	for _, cs := range statuses {
+		switch {
+		case cs.State.Waiting != nil:
+			t.Logf("  %s %s: waiting (reason=%s)", containerType, cs.Name, cs.State.Waiting.Reason)
+		case cs.State.Running != nil:
+			t.Logf("  %s %s: running", containerType, cs.Name)
+		case cs.State.Terminated != nil:
+			t.Logf("  %s %s: terminated (exitCode=%d, reason=%s)",
+				containerType, cs.Name, cs.State.Terminated.ExitCode, cs.State.Terminated.Reason)
+		}
+	}
+}
+
+// WaitForCrashLoopBackOff waits for CrashLoopBackOff in either containerStatuses or
+// initContainerStatuses based on the checkInit flag.
+// When checkInit=false, checks main container statuses; when checkInit=true, checks init containers.
+func WaitForCrashLoopBackOff(
+	ctx context.Context, t *testing.T, client klient.Client,
+	namespace, podNamePrefix string, checkInit bool,
+) *v1.Pod {
+	t.Helper()
+
+	var foundPod *v1.Pod
+
+	containerType := "Container"
+	if checkInit {
+		containerType = "InitContainer"
+	}
+
+	require.Eventually(t, func() bool {
+		var podList v1.PodList
+		if err := client.Resources(namespace).List(ctx, &podList); err != nil {
+			return false
+		}
+
+		for i := range podList.Items {
+			pod := &podList.Items[i]
+
+			// Find pod by prefix (DaemonSet pods have generated suffixes)
+			if !strings.HasPrefix(pod.Name, podNamePrefix) {
+				continue
+			}
+
+			// Select which container statuses to check
+			statuses := pod.Status.ContainerStatuses
+			if checkInit {
+				statuses = pod.Status.InitContainerStatuses
+			}
+
+			if CheckPodCrashLoopBackOff(statuses) {
+				t.Logf("Pod %s %s is in CrashLoopBackOff (phase: %s)",
+					pod.Name, containerType, pod.Status.Phase)
+				foundPod = pod
+
+				return true
+			}
+
+			// Log current state for debugging
+			t.Logf("Pod %s: phase=%s, %sStatuses=%d",
+				pod.Name, pod.Status.Phase, containerType, len(statuses))
+			LogContainerStatuses(t, statuses, containerType)
+		}
+
+		return false
+	}, EventuallyWaitTimeout, WaitInterval,
+		"pod with prefix %s %s did not enter CrashLoopBackOff", podNamePrefix, containerType)
+
+	return foundPod
+}
+
+// WaitForDaemonSetPodRunning waits for a DaemonSet pod to reach Running state with all
+// containers ready. This is useful after fixing a crashing container to verify recovery.
+func WaitForDaemonSetPodRunning(
+	ctx context.Context, t *testing.T, client klient.Client, namespace, dsName, nodeName string,
+) {
+	t.Helper()
+
+	require.Eventually(t, func() bool {
+		pods, err := ListDaemonSetPods(ctx, client, namespace, dsName)
+		if err != nil {
+			t.Logf("Error listing pods: %v", err)
+			return false
+		}
+
+		for _, pod := range pods {
+			// Check if pod is on the expected node
+			if pod.Spec.NodeName != nodeName {
+				continue
+			}
+
+			// Check if pod is Running
+			if pod.Status.Phase != v1.PodRunning {
+				t.Logf("Pod %s phase: %s (waiting for Running)", pod.Name, pod.Status.Phase)
+				return false
+			}
+
+			// Check if all containers are ready
+			allReady := true
+
+			for _, cs := range pod.Status.ContainerStatuses {
+				if !cs.Ready {
+					t.Logf("Pod %s container %s not ready yet", pod.Name, cs.Name)
+
+					allReady = false
+				}
+			}
+
+			if allReady && len(pod.Status.ContainerStatuses) > 0 {
+				t.Logf("Pod %s is Running with all containers ready", pod.Name)
+				return true
+			}
+		}
+
+		return false
+	}, EventuallyWaitTimeout, WaitInterval,
+		"DaemonSet %s pod on node %s did not reach Running state", dsName, nodeName)
+}
